refactor: TuripNavigationBar 위치 이동 및 패키지 경로 수정
